# Improve performance of String.IndexOf(char) and String.LastIndexOf(char)