@bbowyersmyth 2016-05-28
Note if we want to add some loop unrolling here the times are at least half of the original CLR C++ code.
It would help with finding '?' and '#' in a url for example.

           Method |      Median |     StdDev | length |
----------------- |------------ |----------- |------- |
 NewIndexOfUnroll |   1.7680 ns |  0.0698 ns |      1 |
 NewIndexOfUnroll |   3.7536 ns |  0.0524 ns |     12 |
 NewIndexOfUnroll |  11.8368 ns |  0.1271 ns |     50 |
 NewIndexOfUnroll |  21.4534 ns |  0.2648 ns |    100 |
 NewIndexOfUnroll |  53.3115 ns |  0.5707 ns |    200 |
 NewIndexOfUnroll | 119.1522 ns |  0.8786 ns |    500 |

The implementation is in the profiling code.