@jkotas 2016-05-29
Thank you for doing all these perf fixes!

The loop unrolling looks pretty good to me. Would you like to update this PR with it?

Replacing `offset` with incrementing `pCh` may be slightly better - something like:
```
	ReturnIndex3: pCh++;
	ReturnIndex2: pCh++;
	ReturnIndex1: pCh++;
	ReturnIndex:
	return (int)(pCh - pChars);
```