@bbowyersmyth 2016-05-29
Thanks @jkotas I'll replace the PR with unrolling after I give your suggestion a profile.

One thing I've noticed is that incrementing a char*, or dereferencing an offset, is more complex than it needs to be. It compiles to a mul by 2 then an add to the pointer. I haven't found in Roslyn where it determines a mul is needed but it would be good to pre-multiply the constant value seeing pointers are normally used for performance.
It doesn't appear that the JIT combines them as I'm seeing (int*)(byte* + 4) being faster than (int*)(char* + 2) in another function when they should be the same for this.