@bbowyersmyth 2016-05-30
Clarification to the above: `pCh++` comes through as an add but `pCh + 2` comes out as a mul + add.

Your increment suggestion has almost the exact same times when a match is found. Some interesting results when no match is found on length 100. I can't explain that one.

            Method |      Median |    StdDev | length |
------------------ |------------ |---------- |------- |
 NewIndexOfUnrollInc |   1.9456 ns | 0.0560 ns |      1 |
 NewIndexOfUnrollInc |   4.1826 ns | 0.0802 ns |     12 |
 NewIndexOfUnrollInc |  12.3469 ns | 0.1354 ns |     50 |
 NewIndexOfUnrollInc |  31.1959 ns | 0.3343 ns |    100 |
 NewIndexOfUnrollInc |  50.0072 ns | 0.4643 ns |    200 |
 NewIndexOfUnrollInc | 111.9723 ns | 1.1223 ns |    500 |

Even still, going with that as it's half the original time and I find it more readable than what I had.