@manu-silicon 2016-05-30
I'm not sure if it has an impact, but what about which I find easier to read:

```
 while (count >= 4)
{
        if ((*pCh == value) || (*++pCh == value) || (*++pCh == value) || (*++pCh == value)) return (int) (pCh - pChars);
        count -= 4;
}
```