@jkotas 2016-05-30
> I'm not sure if it has an impact

This pattern would have negative impact.