@bbowyersmyth 2016-05-30
Hey Manu, a few things there. The first `if` comparison needs an increment for the next iteration. After fixing that logic and starting pCh at -1 the above code for the 500 length runs at 146ns where as the PR is at 111ns.
Most of the structure of the PR is designed to reduce data dependencies and loop complexity. It is not a structure I would recommend for normal use but this is expected to be performance system code.