@manu-silicon 2016-05-31
@bbowyersmyth Thanks for running the benchmarks and fixing the code :-). I ran it on my end against the .NET Framework. I got similar results. One thing I found out was that the return statement in the loop had a negative impact. For my proposed loop, it went from 165ns with a return vs 145ns with a goto.